Create Write a program to accept 3 integers and find maximum among 3 numbers using functions and pointers.
A pointer is a variable that stores the memory address of another variable. It allows indirect access to the
value stored in that memory location. Pointers are powerful tools in C++ programming and are used for
various purposes, such as dynamic memory allocation, passing parameters to functions by reference.
Write a program to accept 3 integers and find maximum among 3 numbers using functions and pointers.